WIP breakup of octodns.record file
