+++ octodns/record/geo.py
@@ -3,7 +3,11 @@
 #
 
 from logging import getLogger
+import re
 
+from ..equality import EqualityTupleMixin
+from .base import ValuesMixin
+from .change import Update
 from .geo_data import geo_data
 
 
@@ -79,3 +83,100 @@ class GeoCodes(object):
         if province in geo_data['NA']['CA']['provinces']:
             country = 'CA'
         return f'NA-{country}-{province}'
+
+
+class GeoValue(EqualityTupleMixin):
+    geo_re = re.compile(
+        r'^(?P<continent_code>\w\w)(-(?P<country_code>\w\w)'
+        r'(-(?P<subdivision_code>\w\w))?)?$'
+    )
+
+    @classmethod
+    def _validate_geo(cls, code):
+        reasons = []
+        match = cls.geo_re.match(code)
+        if not match:
+            reasons.append(f'invalid geo "{code}"')
+        return reasons
+
+    def __init__(self, geo, values):
+        self.code = geo
+        match = self.geo_re.match(geo)
+        self.continent_code = match.group('continent_code')
+        self.country_code = match.group('country_code')
+        self.subdivision_code = match.group('subdivision_code')
+        self.values = sorted(values)
+
+    @property
+    def parents(self):
+        bits = self.code.split('-')[:-1]
+        while bits:
+            yield '-'.join(bits)
+            bits.pop()
+
+    def _equality_tuple(self):
+        return (
+            self.continent_code,
+            self.country_code,
+            self.subdivision_code,
+            self.values,
+        )
+
+    def __repr__(self):
+        return (
+            f"'Geo {self.continent_code} {self.country_code} "
+            "{self.subdivision_code} {self.values}'"
+        )
+
+
+class _GeoMixin(ValuesMixin):
+    '''
+    Adds GeoDNS support to a record.
+
+    Must be included before `Record`.
+    '''
+
+    @classmethod
+    def validate(cls, name, fqdn, data):
+        reasons = super().validate(name, fqdn, data)
+        try:
+            geo = dict(data['geo'])
+            for code, values in geo.items():
+                reasons.extend(GeoValue._validate_geo(code))
+                reasons.extend(cls._value_type.validate(values, cls._type))
+        except KeyError:
+            pass
+        return reasons
+
+    def __init__(self, zone, name, data, *args, **kwargs):
+        super().__init__(zone, name, data, *args, **kwargs)
+        try:
+            self.geo = dict(data['geo'])
+        except KeyError:
+            self.geo = {}
+        for code, values in self.geo.items():
+            self.geo[code] = GeoValue(code, values)
+
+    def _data(self):
+        ret = super()._data()
+        if self.geo:
+            geo = {}
+            for code, value in self.geo.items():
+                geo[code] = value.values
+            ret['geo'] = geo
+        return ret
+
+    def changes(self, other, target):
+        if target.SUPPORTS_GEO:
+            if self.geo != other.geo:
+                return Update(self, other)
+        return super().changes(other, target)
+
+    def __repr__(self):
+        if self.geo:
+            klass = self.__class__.__name__
+            return (
+                f'<{klass} {self._type} {self.ttl}, {self.decoded_fqdn}, '
+                f'{self.values}, {self.geo}>'
+            )
+        return super().__repr__()
